Refactor: `TagService` 리팩터링
